+++ R/html-helpers.R
@@ -1,13 +1,14 @@
 #' Add a simple table with column names and matching labels
 #'
 #' @param label A string representing the label for the details expansion section.
-#' @param content A named list
-#'
+#' @param content A named list or wide data.frame with 2 rows
+#' @param names a string indicating the name of the two columns inside the details tag
 #' @return HTML text
 #' @export
-gt_label_details <- function(label, content){
+gt_label_details <- function(label, content, names = c("Column", "Description")){
 
   stopifnot("Must be a named list" = length(names(content)) >= 1)
+  stopifnot("'names' must be length 2" = length(names) == 2)
 
   build_content <- function(lab_item, content_item){
     glue::glue(
@@ -23,7 +24,7 @@ gt_label_details <- function(label, content){
     paste0(collapse = "")
 
   c(glue::glue("<details><summary>{label}</summary>"),
-    "<br><table><tr><th>Column</th><th>Description</th>",
+    glue::glue("<br><table><tr><th>{names[1]}</th><th>{names[2]}</th>"),
     fill_content,
     "</table></details>") %>% paste0(collapse = "") %>%
     as.character() %>%